power: charger: fix a bug about charging icon which dose not disappear when charger is gone
When battery is full, it will set a flag to represent this event. The flag is cleared
in charger gone interrupt process. However, when charger is gone, the gone interrupt
has not happen after a long time low probability. This may be a hardware error.
So when the bug happens, fix it in charging status interface and usb online interface.

Change-Id: I3c00276ce361b967a1df8cfaba0c46b9566b8e0f